install: Insist on full tree when mutating with a package.lock
That is, if you have a package-lock but haven't installed anything yet then
we're going to insist that all of the bits be there. This is necessary in order
to be able to cleanly upgrade old package-lock files. This means if you type
`npm install foo` or `npm rm foo` and don't have a `node_modules` but do have
a `package-lock.json` it's gonna install everything from the
`package-lock.json` in addition to making your change.

The one thing we won't do is update your package-lock from your
package.json.  We only do that when you request a full install.

This does not change the behavior if you don't have a lock file.